server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pydantic==2.10.4
slowapi==0.1.9
httpx==0.28.1
PyJWT==2.10.1
